Телеграмм бот по образу и подобию @this_burger_bot готов. Комментарии Java docs на русском добавлены. Удалены все лишние классы.

+++ backend/src/main/java/com/example/backend/repository/OrderedDishRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderedDishRepository extends JpaRepository<OrderedDish, Integer> {
+public interface OrderedDishRepository extends JpaRepository<OrderedDish, Long> {
 }
